167238950-bug(heroku)-Fix deployment issues:
- Change webpack config

[#167238950 Starts]
