+++ webpack/webpack.config.base.js
@@ -1,18 +1,6 @@
 const path = require("path");
-const webpack = require("webpack");
 const htmlWebpackPlugin = require("html-webpack-plugin");
-const dotenv = require("dotenv");
 
-/**
- * get the env variables
- */
-const env = dotenv.config().parsed;
-
-// reduce it to a nice object, the same as before
-const envKeys = Object.keys(env).reduce((prev, next) => {
-  prev[`process.env.${next}`] = JSON.stringify(env[next]);
-  return prev;
-}, {});
 
 module.exports = {
   entry: "./src/index.js",
@@ -53,7 +41,5 @@ module.exports = {
   plugins: [
     new htmlWebpackPlugin({
       template: "./src/index.html"
-    }),
-    new webpack.DefinePlugin(envKeys)
-  ]
+    })  ]
 };
