+++ webpack/webpack.config.prod.js
@@ -3,6 +3,7 @@ const workboxPlugin = require("workbox-webpack-plugin");
 
 const baseConfig = require("./webpack.config.base");
 
+
 module.exports = merge(baseConfig, {
   mode: "production",
   plugins: [
